move panel measure hook
